static globals

+++ enums_typedef.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 enum class exchange {NYSE = 1, NASD, ARCA, BATS, EDGX, EDGA};
 
 enum class side {BUY, SELL, NONE};
 
-typedef int Price;
-typedef int Shares;
-typedef int Timestamp;
-typedef int OrderId;
+typedef int64_t Price;
+typedef int32_t Shares;
+typedef uint64_t Timestamp;
+typedef int32_t OrderId;
